requirements.txt
google-generativeai 
python-dotenv
rarfile
mcp[cli]
pydantic>=2.0.0